refactory to hook
